Bug#25654833 MYSQL CLUSTER 7.5.5 REPLICATION SLAVE SQL THREAD HANGS WITH CREATE TABLE
This is 'best effort' patch for the above problem, which seems to be a regression introduced in 7.5.5. Based on bug description, we *assume* this to be caused bu the 'part 3 of 4' push of patch for Bug#25042101 'SPLIT BINLOG INJECTOR_MUTEX IN TWO, DO REQUIRED CLEANUP':

.........
  Remove thread_yield() in binlog injector code previously put there
  as a temp stopgap in the commit below. This used to be required
  as the injector thread held the injector_mutex > 99% of the time when
  waiting for pollEvents(). That blocked client threads either wanting to
  access the data shared from the injector thread, or needing the injector_mutex
  while waiting for injector_cond to be signaled

  This should not be required anymore, as:

  1) injector_mutex has been splitt in two separate mutexes.
  2) We changed init of the injector_event_mutex from a 'FAST' to a 'SLOW'
     mutex which has better 'farness' properties in the scheduler
.........

The theory is that using the more 'fair' SLOW-mutex variant is not as 'fair' as assumed to be on all OS/VM variants. Thus the 'yields' may be unsafe to remove.

Bug has been hard to reproduce:

 - It has been 'randomly' reproduced using Oracle Linux 7 on top
   of 'Virtualbox'
 - It has than gone away for a couple of days.
 - It has *not* been reporoduced with this patch.

So it should be reasonable to expect that reintroducing the 'thread_yields' should solve the problem.

We cant say for sure though....