Bug #25476474 SIGNAL 11 AT HA_NDBCLUSTER::EXEC_BULK_UPDATE
A bulk update is executed by reading records and executing a
transaction on the set of records. The transaction is started
while reading the set of records. In this case, the transaction
start fails. The transaction execution code then attempts to
dereference the NULL transaction pointer, which causes a segfault.

The mysql update function does not exit on fatal errors. The
approach is to save all errors, continue execution and evaluate
success/failure after the execution completes. With this approach,
it is not guaranteed that the necessary init work is successfully
completed before calling the bulk update executor. So it is
necessary to have more error-handling in the ndb bulk update
executor.

Fixed by adding some error-handling in the ndb bulk update
executor.