WL#8069: Added more debugging information and crash information and also more checks of consistency when reading disk pages to ensure that we find inconsistencies ASAP.
Fixed a bug in deletion of LCP files. The deletion of LCP files relies on
the RESTORABLE_GCI_REP being sent to the BACKUP block as soon as a GCP is
fully completed. However at System Restart no such signals are sent.
Thus we won't delete any LCP files during System Restart. This is bad in
itself, but even worse is that the System Restart hangs in LCP handling
in the PREPARE_DROP state when a table is dropped as part of the
System Restart. Fixed by NDBCNTR sending RESTORABLE_GCI_REP to BACKUP
block with the restorable GCI.