Bug#25741170 SPJ: DON'T ADD FLUSH_AI-PSEUDO_READS WHEN NOT NEEDED
Follow up patch to fix a regression detected by ATR test
./testSpj -n NF_Join T6

In case the SPJ results has to be sent by the routed TRANSID_AI_R
signal the block ref for the 'route' destination refered the
SPJ block instead of the TC block.

It turns out that for scan type SPJ request LQHKEYCONF/REF
is expected by the SPJ for the lookup-leaf treeNode. The
LqhKeyReq::tcBlockref is set up to contain the SPJ block
which such CONF/REF should be sent to.

Unfortunately the same ::tcBlockref field is used in cases where
a TRANSID_AI_R signal has to be used, which in these cases will
refer a SPJ block.

Thus we have to use the FLUSH_AI which contain its own 'routeRef'
block in a SPJ scan-request.

This patch fix the above issue and revert the part of the previous
patch which set up scan requests for eliminating unneeded FLUSH_AI's

This causes us to loose some important oportunities for sending
packed 'short' TRANSID_AI signals. However, this is assumed to be
regained by the upcomming patch for:

bug 25750355 SEND SPJ RESULTS AS 'PACKED' TRANSID_AI WHERE APPLICABLE