Bug #25769683: STRANGE ISSUE WITH ADMIN COMMAND "LIST_OBJECTS"
Description:
An issue is related to xplugin admin command "list_objects". The
command is used to check if given collection exists in a database or
not. The issue is that it happens only on OSX, only if schema name
contains an uppercase letter.

Solution:
The 'list_objects' admin command SQL query has been improved to such
form to take into account a size of letters in names of tables and
schemas in a process of collecting them.

Reviewed-by: Lukasz Kotula <lukasz.kotula@oracle.com>
RB: 15945