BUG#25041396 --LOG-BIN MARKS MANY PERFORMANCE SCHEMA TABLES INVALID
Problem:
When a performance_schema table was dropped or does not
have the expected structure, and later the server was
restarted with binlog enabled, it marked all performance
schema tables encountered after the first error as invalid.

Analysis:
The method decide_log_format would return with an error if
there was any currently active error and the table was not
to be replicated (as it is the case with P_S tables).
This error was being generated when checking the P_S tables
upon restart. However, it is sufficient to print the error
to the log, because the check is not supposed to return
any error.

Fix:
Since checking the performance schema at server start is not
supposed to generate an error, it is safe to clear the error
in case there is one.