BUG#25800025: SQL THREAD ERROR 1778 WHEN GTID SKIPPING PARTIAL TRX
Problem
-------

SQL thread is unable to GTID skip a partial transaction.

Analysis
--------

The SQL thread has a mechanism to rollback the partial transactions left
on the relay log. The mechanism is activated every time the SQL thread
is going to apply a Gtid_log_event and the THD already owns a GTID.

The issue is happening when the transaction being applied is to be GTID
skipped. Because in this case the THD owns no GTID, the rollback
mechanism is not being used.

Fix
---

When applying a Gtid_log_event, the SQL thread will "rollback" the
partial transaction when a transaction that was being skipped was
incomplete in the relay log.