Bug#25636986 Add header files to project sources so XCode IDE can find them when when searching
Currently, the Xcode IDE has a Find->Text->Containing pane that will
search through MySQL project files for text strings. With the existing
Cmake generated Xcode project "MySQL", the header files are not
included in the sources, so they will not be searched.

This may possibly apply to other IDEs as well.

This patch makes sure the header files in the

      sql
      include
      storage

directories and their subdirectories are added to the project sources.

Change-Id: I368e09f954659cb4bdc1e346a9c08c34600c8e00