BUG#25681977 GROUP REPLICATION NODE ON OPC FAILED TO REJOIN AFTER REBOOT
PROBLEM
-------
For many maintenance operations, VMs on OPC need to be restarted, but
after a VM restart, the node fails to rejoin the GR group. The problem
is that the private IP is changed after a VM restart and the remaining
nodes have a cached reference to the previous private IP.

When a node reconnects, a node's description called "server" is reused
if there is any. The server contains a reference to the addrinfo
structure and the current implementation does not check whether the
stored information is still valid or not.

FIX
---
To fix this issue, a new addrinfo structure is always fetched while
adding a new node or connecting to it.