Bug#11766869 NDB_CONFIG TO DUMP COMPLETE CURRENT CONFIGURATION
We need to use query command to get value of current configurations from
the ndb_config tool. There are almost 250 different configuration and
if suppose user require all the configuration parameter, it
is very difficult for the user to call them one by one with query, hence there
arise the need for a new option type "--query-all" which dump all the configuration .

Have added a new option in ndb_config to dump the entire configuration.