Bug#25812494: MAIN.LOCK_SYNC IS ALWAYS SKIPPED
Issue:
------
main.lock_sync has an opt file containing --binlog_format=STATEMENT.
It also contains the have_binlog_format_mixed_or_statement.inc. Therefore,
the valid binlog formats for the test will be 'mixed' and 'statement'.
However, When 'STATEMENT' is encountered, MTR thinks it is not a valid
format due to the difference in case.

Fix:
----
The comparison of binlog formats in the opt file with the valid binlog formats
has been made case insensitive.

Reviewed-by: Pavan Naik <pavan.naik@oracle.com>
RB: 15904