WL #9554: Varlen keys for sorting multibyte String values
Unicode collations are sorted by means of variable-length weight strings.
In the worst case, these weight strings can become very long; for instance, for
utf8mb4_0900_as_cs, we have a strnxfrm_multiply of 24, which is interpreted as
every byte in the input string becoming potentially 24 bytes. In other words, a
VARCHAR(100), which can be up to 400 bytes, gets 9600 bytes allocated for
filesort, even if it just contains a simple 'a' (which is six bytes of weight
plus some level separators). The default set max_sort_length=1024 masks this
problem somewhat by truncating the weight strings, at the cost of incorrect and
unpredictable sorting when sorting strings that actually need long weight
strings.

This worklog aims to introduce variable-length keys when sorting NO PAD collations.
(PAD collations still need to be fixed length, because they are conceptually
extended to infinity.) It builds on the existing semantics for sorting JSON using
variable-length keys; it doesn't try to replace strnxfrm with strnncollsp,
which would also be an interesting avenue, but can happen in a later worklog.

Benchmark results are good: For the distinct_ranges sysbench test
(using varchar keys), we win about 12% on a single core of Skylake 3.4 GHz.
For a test of sorting 300000 first names in a VARCHAR(100) COLLATE utf8mb4_0900_as_cs,
setting max_sort_length=65536 (so that both sides actually sort based on all
their data, as opposed to truncating the key and sacrificing correctness),
it is about five times as fast, due to vastly less spilling to external storage
(77 versus 11075 chunks). This is even though external storage in this case is
RAM, not disk -- if it would be actually hitting the SSD, the difference would
be even more dramatic.

Change-Id: I4704ffe04e2b35db47e5c1b98160f6a4731157ad