Bug #25361251: INCORRECT BEHAVIOR WITH INSERT ON DUPLICATE                KEY IN SP
Issue 1:
--------
The VALUES part of the ON DUPLICATE KEY UPDATE can refer to
the column list of an INSERT statement. When a BLOB value
is used in such a clause as part of an INSERT statement in
stored procedure, this can result in incorrect behavior.

BLOB values require some special handling because of
problems like the one mentioned in Bug#13901905. This is
done using the "insert_update_values_map" member of LEX.
Field_blob values are added to this map during
Item_insert_value::fix_fields. But the cleaning up takes
place at the end of the statement in lex_end.

This works fine for conventional execution statements. But
for statements in the body of an SP, lex_end is called only
once per statement. This is irrespective of whether the
statement is executed in a loop. This can result in
incorrect behavior because "insert_update_values_map" needs
to be cleaned up after statement execute.


Solution 1:
------------
The solution is to make sure that lex->clear_values_map is
called at the end of statement execution.


Issue 2 (from Bug #24538207):
-----------------------------
The problem here is similar to the one mentioned above. The
INSERT is a prepared statement. Each PREPARE also creates a
new "insert_update_values_map" for handling the BLOB values.
Only lex_end cleans up here too.

This results in the same problem as above.

Solution 2:
------------
The "EXECUTE ..." part will be fixed by Solution 1. But for
the "PREPARE ..." statements, lex->clear_values_map will be
added at the end of "Prepared_statement::prepare".