Bug #25830247   NDB : CPUSET LOCKING DOES NOT REUSE CPUSET STRUCTURES
There is code spread across NdbLockCpuUtil.c and NdbThread.c for managing a set of cpuset arrays.  The idea is that a minimal set of unique cpusets will be maintained.  Cpusets can be reused by different threads, with reference counting used to avoid leaks.

However :
 - The code to reuse a cpuset does not work as the cpuset content is not updated when a cpuset is created, so the search cannot work
 - The code to release a cpuset does not work as the cpuset id of a thread is not read before it is nulled, leaking the cpuset.

Probably these bugs are not too problematic for static cpuset assignments, but they cause problems when the number of unique cpusets is high, or if cpuset locking is changed at runtime.