Bug#25851801 7.6.2(DMR2):: COMPLETE CLUSTER CRASHED DURING UNIQUE KEY CREATION ...
Demonstration of bug.

Use error injection to simulate failure of DynArr256Pool::seize() that in
turn make Dbacc::setPagePtr fail.

The later function registers the page number to an fragment.

Later when fragment pages are to be released DynArr256::truncate() are
called which assumes at least one directory page have been seized if
set() been called.

This is not the case.