Bug#25851801 7.6.2(DMR2):: COMPLETE CLUSTER CRASHED DURING UNIQUE KEY CREATION ...
DynArr256::truncate() now succeeds if there are no pages allocated in array.

The test case with the above fix now brought up another bug there two calls to
Dbacc::releasePage() is missing when pages was not possible to add to fragments
page directory due to out of memory when allocating directory page.

This made some checks failed and the dbacc page was leaked.

The missing calls to releasePage() were added.