Bug #24323288: MAIN.MYSQLBINLOG_DEBUG FAILS WITH A LEAKSANITIZER ERROR
Issue:
------
On a system with AddressSanitizer support that includes Leaksanitizer,
the test mysqlbinlog_debug fails.test when compiled with -DWITH_ASAN and
-DCMAKE_BUILD_TYPE=Debug option.

Cause:
------
When the events are read from remote server using mysqlbinlog utility,
in two scenarios the memory clean up is not done. Those scenarios are.

1) When fake Rotate_event occur. Places where
   Rotate_event->when.tv_sec == 0
2) When we are using raw mode with mysqlbinlog, then for those cases
   where the log file could not be created or write had some issues.

Solution for 5.6 and 5.7:
-------------------------
5.6:
===
Scenario (1) will be fixed by backport of Bug#21697461 from mysql-5.7,
which is to create a method that will handle the memory clean up and
call this method from all the suspected places.

Scenario (2) will be an extension to Bug#21697461, for cases discussed
above in the *cause* section.
5.7:
===
As we already have Bug#21697461 in mysql-5.7, scenario (1) need not be
addressed separately.
For scenario (2) same fix as above, extension of Bug#21697461.

Solution for trunk:
-------------------
Implemented a different solution(removed Bug#21697461) to fix scenario
(1) and (2).

- A new solution is implemented to make the area around the
- Manage the clean up of temp_buf through m_free_temp_buf_in_destructor
  variable added in Log_event class.
- Created a new RAII class Destroy_log_event_guard to handle memory of
  Log_event object created in the dump_remote_log_entries method.
- Extended mysqlbinlog_debug.test to test the changes done.

A new solution is implemented thinking that any modification done to
the area around the code causing memory leaks at present have lesser number
of memory issues in future.