Bug #21960004
SETTING MAX_ROWS=0 IS NOT THE SAME AS NEVER SETTING MAX_ROWS

An inplace ALTER TABLE .. MAX_ROWS query intentionally updates only the
mysqld metadata and not the NDB dictionary. And the REORGANIZE
PARTITION query first checks the NDB dictionaries max_rows value to
decide whether to reorganize or not. Due to this outdated value in the
NDB dictionary, REORGANIZE is denied even if the max_rows is set back
to 0.

This patch fixes this issue by using the max_row value provided by MySQL
instead of the NDB dictionary value to make any decisions.

Note : This fix is only for versions 7.3 and 7.4.
       Versions 7.5 and above have a slightly different patch.