Bug#25775881 - INNODB ASSERTION DURING INITIALIZE IF FAILED TO CREATE TABLESPACES.OPEN.1
Problem:
--------
The problem happens only when mysqld process is run as root user and
creates datadir as mysql user. When InnoDB creates tablespaces.open.(1|2)
files using absolute path to fopen(), file creation fails with permission
denied error. Since this file cannot be opened for writing, a fatal error
is thrown.

Fix:
----
When creating the tablespaces.open.(1|2) files, an absoulte path is formed.
This is not necessary. Just create the path based on srv_log_group_home_dir
and the tablespace.open.(1|2) file.

Reviewed-By: Jimmy Yang <jimmy.yang@oracle.com>
RB: 16012