Bug #24753281: REQUEST FOR SLAVE_SKIP_ERRORS = ER_INCONSISTENT_ERROR
Problem:
--------
Replication applier thread returns 3002, ER_INCONSISTENT_ERROR, when there's
difference between expected error number and actual error number on slave.

So when ER_INCONSISTENT_ERROR is passed as an input to slave-skip-errors,
the replication should continue despite having different expected error and
actual error on slave.

But in current code, the slave is not able to ignore ER_INCONSISTENT_ERROR.

Fix:
----
Check if ER_INCONSISTENT_ERROR is passed in the list of errors to be
skipped. If yes then do not throw error, just log this info with the mention
of actual and expected error  as a note in the slave's error log.