Bug #23141739 : ERROR 240 IS NOT DEFINED
The error 240 is thrown when there is a mismatch between the fk
trigger columns and the values supplied to them during the trigger
execution. This mismatch is actually a side effect of the Bug#23068914
due to which the distributed TC and LDM threads end up with different
trigger ids for the same FK trigger after a restart. This base bug has
already been fixed.

This patch adds an error description for the error 240 and extends the
MTR testcase of Bug#23068914 to cover this issue.