Bug #25844166 NDB_CONFIG TO PRINT CONFIG DIFFERENT FROM DEFAULT
There are more than 256 configurations in cluster. So, when we like to find
the configuration that are changed during run time it is difficult to find
them, hence have added an option(--diff_default) to ndb_config to print
only the configurations that are different from the default configuration
of the node.