Bug#25252679 : ENABLE MAIN.MYSQL_UPGRADE.TEST
mysql.ndb_binlog_index table is created in mysql schema
when server is built with NDB cluster. This result in an extra
entry in output of mysql_upgrade. Hence, tests that execute
with both and without ndbcluster, fails with 'result content
mismatch' when mysql_upgrade is executed.

Update the tests to filter out results related to the
ndb_binlog_index table.