BUG#25476479:  MYSQLD FILLS UP ERROR LOGS WITH [ERROR]                ERROR IN ACCEPT: BAD FILE DESCRIPTOR
DESCRIPTION:
============
When mysqld is secured with TCP wrappers and the files
hosts.allow and hosts.deny are configured to restrict access
from an IP address, very large error log files are generated
if we try to connect to the server via that IP address.
This runs indefinitely resulting in ever growing error log
files and puts load on the machine. The server continues to
serve authorized connections although slowly.

ANALYSIS:
=========
When the listening socket is accepting connections, if a
connection was refused by TCP wrappers, we were closing the
listening socket which led to connect socket being polled
continuously, leading to indefinite number of errors being
logged to the error log.

FIX:
====
We are now closing the connect socket only, thus avoiding
its polling and also allowing the listening socket to
accept other connections.