Bug#25945568: MYSQL 5.7 FAILS TO BUILD WITH MUSL LIBC               DUE TO NOT FINDING SIGEV_THREAD_ID
The problem was that posix_timers relied on non-posix standard
features which made the code not compile on Alpine Linux.
This patch rewrites the code to use standard posix. This means
that the ~same code can also be used on Solaris.

Also fix various minor other issues for compiling on Alpine.