BUG#11766444 : MTR PRINTS WRONG FILE AND LINE NUMBER WHEN TESTS FAIL
Description :
=============
When a test case fails, MTR sometimes prints the wrong line number.
This happens if a test case fails inside a while loop and it is not
the first iteration of the loop.

E.g:
----
--let $i = 2
while ($i)
{
  dec $i;
  if (!$i)
  {
    execute this invalid query;
  }
}

When executing the above test case, MTR fails with the following
message:

mysqltest: At line 9: query 'execute this invalid query' failed: 1064:
You have an error in your SQL syntax; check the manual that
corresponds to your MySQL server version for the right syntax to use
near 'invalid query' at line 1

Note that the error was on line 7 but the test reports line 9.

Issue :
=======
The variable 'start_lineno' keeping track of the current line number
in a test is not updated after the first iteration and hence printing
the line number of the last line of the loop.

Fix :
=====
Introduced a new variable 'lineno' in 'st_command' structure to keep
track of the the line number for each command. During the iteration
after the first one, use this variable to update the 'start_lineno'
variable to the current line number.

Reviewed-by: Deepa Dixit <deepa.dixit@oracle.com>
Reviewed-by: Srikanth B R <srikanth.b.r@oracle.com>
RB: 16015