Bug #25333414 CLUSTER FAILS TO RESTART AFTER ERROR 4239 'TRIGGER WITH GIVEN NAME ALREADY EXIST
The data node fails to restart because the NoOfFragmentLogParts
config parameter is set to 16 even though only one LDM is running.
All 16 log parts are assigned to the same LDM for redo log replay.
This results in page exhaustion during redo log replay and causes
a crash.

Added a fix to to detect a config with more than 4 redo log parts
per LDM, and reject it as an invalid config.