wl#10234: Use multiple fragments in SCAN_FRAGREQ to SPJ
Refactor patch preparing the ground for wl#10234:

There are currently two SCAN type treenodes in the SPJ block:

- The QN_SCAN_FRAG (scanFrag* methods) scans a table,
optionally through an index. It is the only scan-treeNode
of the two which could be the 'root' of a scan type SPJ query.
However, it is limited to scaning a single fragment.

- The QN_SCAN_INDEX (scanIndex* methods) scans an child-treeNode
which is linked to its parent treeNode trough a range equality
on a (sub set of) its non unique indexed columns.
Contraery to the QN_SCAN_FRAG, it is able to scan *all* fragments
belonging to its table across the data nodes.
However a QN_SCAN_INDEX can currently not be the
'root' in a tree of SPJ operations.

As WL#10234 will need even the root node to be able to handle
scan across multiple fragments, some of the above limitations
has to be lifted, Thus this patch:

 - Extend the SPJ scanIndex pseudo class to also be able to
   act as the root of a scan operation.
 - Dbspj::build() convert an incomming QN_SCAN_FRAG node
   to a SCAN_INDEX node.
 - Introduce a 'deprecation assert' in all existing scanFrag<foo>
   methods: They are now replaced with the scanIndex methods.
   (To be removed in later patch)

This leaves the scanFrag* methods unused, but still present
in the codebase.

The following patches will eventually remove the scanFrag*
methods and replace all its usage with the extendes scanIndex*
methods.