BUG#23699215 : MTR TESTS FAILING ON WINDOWS IF VAR DIR HAS '\' INSTEAD OF ' /' PATH SEPARATOR
Issue:
=======
On windows, backslashes in the file name argument to "load data
infile" statement should be specified either as forward slashes or
doubled backslashes. If vardir path contains backslashes,
"check-warnings.test" will fail with parallel > 1, because the
path to the error log file is calcualted using vardir path and this
path is used with "load data infile" statement in that test.

Fix:
====
Replace path seperator '\' with '/' on windows.

Reviewed-by: Deepa Dixit <deepa.dixit@oracle.com>
Reviewed-by: Srikanth B R <srikanth.b.r@oracle.com>
RB: 16067