Bug#25287633:GCOLS: INCORRECT BEHAVIOR WITH CHARSET CHANGES
Issue:
------
Virtual generated column functions that contain certain
string functions can cause problems when there is a change
in charset.

Steps:
1) Create table t1 with virtual generated column containing
string functions that generate strings in the connection
character set.

2) SELECT * FROM t1;
This step creates the item tree for the generated column.
While creating the objects, the table's memroot is used.
The flag "permanent_changes_completed" is set to true to
indicate that no further permanent changes will take place
in the item tree.

3) SET NAMES utf8;
Changes the connection's default charset to utf8.

4) SELECT * FROM t1;
While re-fixing the generated column items
(TABLE::refix_gc_items), it is noticed that the default
charset of the result of some string functions doesn't
match with other parameters. So a new Item object is
created with the correct charset. This new object now
replaces the one present in the item tree (created in
step 2).

But the new object is created on the statement's
mem_root and hence will be destroyed at the end of the
statement. Thus leaving the generated column's item tree
with a dangling pointer.

5) SELECT * FROM t1;
While going through the generated column's item tree,
dangling pointer strikes.

The root cause here is that a conversion due to charsets is
required and a new Item creation is required in step 4.
This should be avoided.

Solution:
--------
In the resolve_type phase, some string functions assign the
current charset to the Item object. This can return strings
in the connection character set.

Use the table's charset (mentioned in the DDL) while doing
fix_fields. Before fix_fields is called on the generated
column's expression, switch to the table's charset. This
will side-step any problem with aggregating charsets since
it is independent of any charset changes made by the user
after the DDL.