BUG#25793366: GR NODE IS IN RECOVERING STATE IF BINLOG_CHECKSUM CONFIGURED ON RUNNING SERVER
When node configured in GR setup with binlog_checksum='NONE' on a running
server, then start group replication return success but node is in RECOVERING
mode forever and server shutdown hangs at this moment. Only option is kill the
server.

The Continuation object used on the plugin applier pipeline to wait for the
applications of events was the source of the issue.  Multiple invocations of
the waiting for execution method when there was a previous error lead to a
stuck pipeline.

As a solution the waiting method no longer waits when there is a known previous
error.