Bug#18513130: STRANGE INTERACTION OF SQL_MODE=ANSI AND PARTITION BY TIMESTAMP
The problem was that the presence of the ANSI_QUOTES SQL mode
influenced how partitioning expressions were stored and
how they were parsed on table definition load.

This made it possible to successfully create a partitioned table
that would result in a parsing error on later retrieval if the
SQL mode changed in the meantime.

This patch fixes the problem by temporarily turning off
ANSI_QUOTES when storing partitioning expressions in the
data dictionary and when they are printed using SHOW CREATE TABLE.
This means that partitioning expressions are no longer stored
exactly as given by the user, but rather using Item::print()
after the expression has been parsed. This is similar to how
view expressions are handled.

This patch also enforces that (sub-)partition expressions
can be maximum 2048 characters long. Before this fix, the
server would assert if a longer expression was used.