wl#10234
Introduce the 'MultiFragFlag' in the ScanTab signal
to indicate that TC should send 'MultiFrag' ScanFragReq.

Also enhance the ScanFragReq signal with the same
'MultiFragFlag' indicator. If set, a list of Fragments
to scan will be sent as the last 'section' in the signal,
after the AttrInfo and KeyInfo (if present)