Bug#25806659 - NDBREQUIRE(LOGFILEPTR.P->FILEREF != RNIL) FAILS, FOR NEXT REDO LOG FILE NOT OPEN
At the last write on a redo log file n, a command to open file n+2 is
issued. At this time, it is expected that file n+1 is already opened
for next writes. However, this may not be the case for a tardy disk,
causing node crash.

This fix will add handling an unopened file gracefully by introducing
waits.